🦄 refactor(all):

+++ 9_排序和搜索/二分/有序数组中是否存在一个数小于等于某个值.py
@@ -3,7 +3,7 @@ from typing import List
 from functools import partial
 
 
-def queryExist(nums: List[float], query: int) -> bool:
+def queryExist(nums: List[int], query: int) -> bool:
     if not nums:
         return False
     index = bisect_right(nums, query)
